Provide `useStore` React hook to sync stateless component with store state
This change provides a `useStore` React hook which can be used in
stateless components to consume a store's state. Its usage looks like
this:

```
const ShowRefreshInterval = () => {
  const { enabled, interval } = useStore(RefreshStore);
  return enabled ? `Refreshing every ${interval}ms` : 'Refresh disabled.';
};
```

It also uses a function mapping the store's state before returning it,
if provided:

```
const ShowRefreshInterval = () => {
  const interval = useStore(RefreshStore, ({ interval }) => interval;
  return `Refreshing every ${interval}ms`;
};
```

`useStore` is properly typed, so if the store is of type `Store<V>`
where `V` is the type of the state it provides, the props mapper
function and the value returned by `useStore` will be properly typed as
well.